+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# (2022-04-22)
+# (2022-04-23)
 
 ### Bug Fixes
 
@@ -254,6 +254,7 @@
 - **graph/leetcode:** add allPathsSourceTarget ([ba33a1e](https://github.com/air-supply94/js-algorithm/commit/ba33a1ec08e4df370da5d830d6bcecd2c18320d4))
 - **graph/leetcode:** add canFinish findOrder ([368b7d3](https://github.com/air-supply94/js-algorithm/commit/368b7d3ca2fcca01df11c449d70fa4149835970c))
 - **greedy:** add canCompleteCircuit ([1a41837](https://github.com/air-supply94/js-algorithm/commit/1a41837c37ce3f8f837f9bb1e53926c26a942bee))
+- **greedy:** add getSmallestString ([904685a](https://github.com/air-supply94/js-algorithm/commit/904685a5092bb3cb8baf1b8e04f0f0a2f93269b2))
 - **greedy:** add jumpGameFast ([680996f](https://github.com/air-supply94/js-algorithm/commit/680996fea200dd5921b18229153f313d7fa5641d))
 - **greedy:** 更新实现逻辑 ([50fd098](https://github.com/air-supply94/js-algorithm/commit/50fd0985a75881f13d4fcb3e31a3678a78a2619b))
 - **hammingDistance:** add ([7540dac](https://github.com/air-supply94/js-algorithm/commit/7540dac75eec4d659f7faeab219084d5b6447dca))
@@ -287,7 +288,7 @@
 - **math:** update countSetBits(增加负数) ([8371770](https://github.com/air-supply94/js-algorithm/commit/837177086d7e2274d2cee79e93a07a7cad317692))
 - **math:** update divingBoard ([d990253](https://github.com/air-supply94/js-algorithm/commit/d990253df88619a5a1091abccd8e64311cc3311e))
 - **math:** update fastPowering ([c193d7a](https://github.com/air-supply94/js-algorithm/commit/c193d7a343827c35ab52fbe25b19882b849c3393))
-- **math:** 更新实现逻辑 ([56ea361](https://github.com/air-supply94/js-algorithm/commit/56ea361e214c8e5ce60dd5a3d335f78495ab1b88))
+- **math:** 更新实现逻辑 ([bdd8d3f](https://github.com/air-supply94/js-algorithm/commit/bdd8d3f1be2bd86d5b03833ad78b7e0cf384bbd7))
 - **math/bits:** add ([8bde116](https://github.com/air-supply94/js-algorithm/commit/8bde116141bc3a86625a2a3a44cdaf79ec6f535e))
 - **math/bits:** test ([3de3260](https://github.com/air-supply94/js-algorithm/commit/3de32608ced958befed76d42b1e8052c2fa6f6cc))
 - **mergeSort:** 更新 mergeSortedArrays concat 方式 ([8baf586](https://github.com/air-supply94/js-algorithm/commit/8baf5861287c1e603d619609628544db119c6781))